+++ projects/plugins/jetpack/_inc/client/components/install-button/index.jsx
@@ -1,6 +1,6 @@
 import { Spinner } from '@wordpress/components';
-import Button from 'components/button';
 import React from 'react';
+import Button from 'components/button';
 import './style.scss';
 
 const InstallButton = props => {