+++ projects/plugins/jetpack/_inc/client/components/plans/plan-icon/index.jsx
@@ -1,3 +1,5 @@
+import PropTypes from 'prop-types';
+import React, { Component } from 'react';
 import { imagePath } from 'constants/urls';
 import {
 	PLAN_FREE,
@@ -108,8 +110,6 @@ import {
 	PLAN_JETPACK_GROWTH_MONTHLY,
 	FEATURE_JETPACK_CRM,
 } from 'lib/plans/constants';
-import PropTypes from 'prop-types';
-import React, { Component } from 'react';
 
 import './style.scss';