+++ projects/js-packages/shared-extension-utils/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.15.20] - 2024-12-02
+### Fixed
+- Fix wrong output type on use-module-status hook. [#40364]
+
 ## [0.15.19] - 2024-11-26
 ### Changed
 - Update dependencies. [#40231]
@@ -494,6 +498,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Core: prepare utility for release
 
+[0.15.20]: https://github.com/Automattic/jetpack-shared-extension-utils/compare/0.15.19...0.15.20
 [0.15.19]: https://github.com/Automattic/jetpack-shared-extension-utils/compare/0.15.18...0.15.19
 [0.15.18]: https://github.com/Automattic/jetpack-shared-extension-utils/compare/0.15.17...0.15.18
 [0.15.17]: https://github.com/Automattic/jetpack-shared-extension-utils/compare/0.15.16...0.15.17